refactor(persona): relocate canonical modules
