+++ pyproject.toml
@@ -44,6 +44,7 @@ py-modules = [
     "local_memory",
     "local_server",
     "latentsync_reply",
+    "mem0_capability_install",
     "mem0_memory",
     "mem0_embedding_install",
     "memory_isolation_case01",
@@ -51,6 +52,7 @@ py-modules = [
     "memory_port",
     "memory_prompt",
     "music_reply",
+    "original_client_capability_api",
     "original_client_companion_api",
     "original_client_companion_backend",
     "original_client_companion_mutation_api",
@@ -58,10 +60,13 @@ py-modules = [
     "original_client_letter_contract",
     "original_client_media_http",
     "original_client_server",
+    "original_client_setup_api",
     "original_client_settings_ui",
+    "original_client_update_api",
     "original_client_video_capability_api",
     "patch_companion_settings",
     "patch_feapp",
+    "patch_webplayer",
     "persona_loader",
     "persona_assembly",
     "persona_provider",
